Create a new class Board and put all game board methods in it
